requirements.txt
asgiref==3.6.0
autopep8==2.0.2
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.1
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
psycopg==3.1.9
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pydotplus==2.0.2
PyJWT==1.7.1
pyparsing==3.0.9
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
sqlparse==0.4.4
typing_extensions==4.5.0
urllib3==2.0.2
